Fix CUDA compile errors

+++ gpu/src/cuda/shaders/utils.cu
@@ -1,5 +1,5 @@
 /// Reverses the `log2(size)` first bits of `i`
-__device__ uint reverse_index(uint i, usize size)
+__device__ uint reverse_index(uint i, uint size)
 {
     if (size == 1)
     { // TODO: replace this statement with an alternative solution.
